+++ src/pyoload/__init__.py
@@ -13,7 +13,7 @@ adds support for:
 :Authors:
   ken-morel
 
-:Version: 2.0.0
+:Version: 2.0.1
 :Dedication: To the pythonista.
 """